判断条码文件是否存在，增加需要删除条码文件的提示

+++ lib/eric_tools/barby_tools.rb
@@ -1,18 +1,20 @@
 module BarbyTools
 
   #生成的图片都比较小,官方的对png控制似乎无效
+  # ------会生成png文件, 一般每个大小不超过1kb, 但是文件需要自己去清除--------
 
-
-  # 生成条码, 不支持utf-8, 比如中文
+  # 生成条码, 不支持utf-8, 比如中文.
   # 支持参数:
   #   content, 要转换的字符串
   #   file_path, 要保存的文件路径(png)
   def self.create_barcode options
     barcode = Barby::Code128B.new(options[:content])
     blob = Barby::PngOutputter.new(barcode).to_png(:height => 20, :margin => 5)
-    File.open(options[:file_path], 'wb'){|f|
-      f.write blob.force_encoding("ISO-8859-1")   #在rails环境中,如果不encode,也会报错
-    }
+    unless File.exist? options[:file_path]
+      File.open(options[:file_path], 'wb'){|f|
+        f.write blob.force_encoding("ISO-8859-1")   #在rails环境中,如果不encode,会报错
+      }
+    end
   end
 
   # 生成二维码. 支持中文
@@ -22,9 +24,11 @@ module BarbyTools
   def self.create_qrcode options
     qrcode = Barby::QrCode.new(options[:content])
     blob = Barby::PngOutputter.new(qrcode).to_png(:height => 20, :margin => 5)
-    File.open(options[:file_path], 'w'){|f|
-      f.write blob.force_encoding("ISO-8859-1")
-    }
+    unless File.exist? options[:file_path]
+      File.open(options[:file_path], 'wb'){|f|
+        f.write blob.force_encoding("ISO-8859-1")
+      }
+    end
   end
 
 end
